Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
boto3==1.34.0
sqlalchemy==2.0.0
psycopg[binary]==3.1.18
greenlet==2.0.2
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
mangum==0.16.0
